fix: shortcut key for grid view undo redo
