+++ apps/nextjs-app/src/features/app/blocks/view/tool-bar/components/UndoRedoButtons.tsx
@@ -1,92 +1,20 @@
 import { Undo2, Redo2 } from '@teable/icons';
-import { redo, undo } from '@teable/openapi';
-import { useTableId, useTablePermission } from '@teable/sdk/hooks';
+import { useTablePermission, useUndoRedo } from '@teable/sdk/hooks';
 import { Button } from '@teable/ui-lib/shadcn';
-import { toast } from '@teable/ui-lib/shadcn/ui/sonner';
-import { useCallback } from 'react';
 import { useHotkeys } from 'react-hotkeys-hook';
 
 export const UndoRedoButtons = () => {
   const permission = useTablePermission();
-  const tableId = useTableId();
 
-  const performUndo = useCallback(async () => {
-    if (!tableId) {
-      toast('nothing to undo');
-      return;
-    }
+  const { undo, redo } = useUndoRedo();
 
-    toast.promise(
-      async () => {
-        const res = await undo(tableId);
-        if (res.data.status === 'fulfilled') {
-          return 'undo success';
-        }
-        if (res.data.status === 'empty') {
-          return 'nothing to undo';
-        }
-        throw new Error(res.data.errorMessage);
-      },
-      {
-        duration: 1500,
-        loading: 'undoing...',
-        success: (message) => message,
-        error: (e) => {
-          return `undo failed: ${(e as { message: string }).message}`;
-        },
-      }
-    );
-  }, [tableId]);
+  useHotkeys(`mod+z`, () => undo(), {
+    preventDefault: true,
+  });
 
-  const performRedo = useCallback(async () => {
-    if (!tableId) {
-      toast('nothing to redo');
-      return;
-    }
-
-    toast.promise(
-      async () => {
-        const res = await redo(tableId);
-        if (res.data.status === 'fulfilled') {
-          return 'redo success';
-        }
-        if (res.data.status === 'empty') {
-          return 'nothing to redo';
-        }
-        throw new Error(res.data.errorMessage);
-      },
-      {
-        duration: 1500,
-        loading: 'redoing...',
-        success: (message) => message,
-        error: (e) => {
-          return `redo failed: ${(e as { message: string }).message}`;
-        },
-      }
-    );
-  }, [tableId]);
-
-  useHotkeys(
-    `mod+z`,
-    async () => {
-      await performUndo();
-    },
-    {
-      preventDefault: true,
-      enableOnFormTags: ['input', 'select', 'textarea'],
-    }
-  );
-
-  useHotkeys(
-    [`mod+shift+z`, `mod+y`],
-    async () => {
-      await performRedo();
-    },
-    {
-      preventDefault: true,
-      enableOnFormTags: ['input', 'select', 'textarea'],
-    }
-  );
+  useHotkeys([`mod+shift+z`, `mod+y`], () => redo(), {
+    preventDefault: true,
+  });
 
   return (
     <>
@@ -95,7 +23,7 @@ export const UndoRedoButtons = () => {
         size={'xs'}
         variant={'ghost'}
         disabled={!permission['table|read']}
-        onClick={performUndo}
+        onClick={undo}
       >
         <Undo2 className="size-4" />
       </Button>
@@ -104,7 +32,7 @@ export const UndoRedoButtons = () => {
         size={'xs'}
         variant={'ghost'}
         disabled={!permission['table|read']}
-        onClick={performRedo}
+        onClick={redo}
       >
         <Redo2 className="size-4" />
       </Button>
